fix datastore long bug

+++ backend/src/main/java/com/google/models/Anomaly.java
@@ -22,6 +22,7 @@ import com.google.appengine.api.datastore.EmbeddedEntity;
 import java.util.Map;
 import java.util.HashMap;
 import java.time.format.DateTimeParseException;
+import java.lang.Number;
 
 /** Store anomaly-related data. */
 public final class Anomaly {
@@ -110,8 +111,9 @@ public final class Anomaly {
     anomalyEntity.setProperty(DIMENSION_NAME_PROPERTY, dimensionName);
 
     EmbeddedEntity dataPointsEntity = new EmbeddedEntity();
+    // Datastore stores numbers as long, so to metricValue should be cast to a long. 
     dataPoints.forEach((timestamp, metricValue) -> 
-        dataPointsEntity.setProperty(timestamp.toString(), metricValue.getValue()));
+        dataPointsEntity.setProperty(timestamp.toString(), (long) metricValue.getValue()));
     anomalyEntity.setProperty(DATA_POINTS_PROPERTY, dataPointsEntity);
 
     return anomalyEntity;
@@ -132,7 +134,7 @@ public final class Anomaly {
 
     if (dataPointsEE != null) {
       for (String key : dataPointsEE.getProperties().keySet()) {
-        dataPointsMap.put(new Timestamp(key), new MetricValue((int) dataPointsEE.getProperty(key)));
+        dataPointsMap.put(new Timestamp(key), new MetricValue(((Long) dataPointsEE.getProperty(key))));
       }
     }
